@@ -11,7 +11,7 @@
     "flux": "^2.0.0",
     "keymirror": "~0.1.0",
     "object-assign": "^1.0.0",
-    "react": "^0.12.0",
+    "react": "^0.13.3",
     "director": "^1.2.7",
     "moment": "^2.9.0",
     "moment-timezone": "^0.3.0",
